student-attendance-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
